backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
pymysql==1.1.0
cryptography==41.0.0
Werkzeug==2.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
